% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{accuracy}
\alias{accuracy}
\title{Calculate map accuracy and standard error}
\usage{
accuracy(err_mat, areas, calculate_area = TRUE)
}
\arguments{
\item{err_mat}{The error matrix, with reference sample and map sample in rows}

\item{areas}{The mapped number of pixels or area of each class (see details)}

\item{calculate_areas}{Whether to produce adjusted area estimates or not.}
}
\value{
A list containing the error matrix, the matrix containing the
estimated proportions of area for each cell, a matrix containing the
accuracy measures, and, optionally, a matrix containing the adjusted areas
estimates for each class with standard errors.
}
\description{
Calculate map accuracy and standard error
}
\details{
Follows the methods provided by Olofsson et al (2014a), producing
estimate of User's, Producer's, and Overall accuracy with standard errors for
each. The outputs can be used to estimate the areas and standard errors of
each class in the `area_estimator` function. If pixel counts are provided to
the areas argument, the resulting area estimates and their standard
errors are returned in pixels, and will require conversion. Note that there
are small differences in outputs here compared to published values due to
rounding issues.
}
\examples{
# Example 1: Table 4 in Olofsson et al 2014a
err_df <- data.frame(class = paste(1:3), r1 = c(97, 3, 2), r2 = c(0, 279, 1),
                     r3 = c(3, 18, 97))
areas <- c(22353, 1122543, 610228)
acc_mat <- accuracy(err_df[, 2:4], areas)
acc_mat  # slight different areas and P

# Example 2: Table 8 in Olofsson et al (2014b)
err_mat <- cbind(class = 1:4, r1 = c(66, 0, 1, 2), r2 = c(0, 55, 0, 1),
                 r3 = c(5, 8, 153, 9), r4 = c(4, 12, 11, 313))
areas <- c(200000, 150000, 3200000, 6450000)
accuracy(err_mat = err_mat[, 2:5], areas = areas)
areas <- (c(200000, 150000, 3200000, 6450000) * 30^2) / 10000
accuracy(err_mat = err_mat[, 2:5], areas = areas)

# Example 3. Table 6 in Olofsson et al (2014a). Areas from Table 2 in
Olofsson et al (2011)
err_mat <- cbind(class = 1:3, logging = c(127, 2, 0),
                 forest = c(66, 322, 15), nonforest = c(54, 17, 540))
areas <- c(logging = 154159, forest = 6846562, nonforest = 16178659)
accuracy(err_mat[, 2:4], areas)
}
\references{
Olofsson, P., Foody, G.M., Stehman, S.V. & Woodcock, C.E. (2014a) Making
better use of accuracy data in land change studies: Estimating accuracy and
area and quantifying uncertainty using stratified estimation. Remote Sensing
of Environment, 129, 122–131.

Olofsson, P., Foody, G.M., Herold, M., Stehman, S.V.,
Woodcock, C.E. & Wulder, M.A. (2014b) Good practices for estimating area and
assessing accuracy of land change. Remote Sensing of Environment, 148, 42–57.

Olofsson, P., Kuemmerle, T., Griffiths, P., Knorn, J., Baccini, A.,
Gancz, V., Blujdea, V., Houghton, R.A., Abrudan, I.V. & Woodcock, C.E. (2011)
Carbon implications of forest restitution in post-socialist Romania.
Environmental Research Letters, 6, 045202.

Stehman, S.V. & Foody, G.M. (2019) Key issues in rigorous accuracy assessment
of land cover products. Remote Sensing of Environment, 231, 111199.
}
