% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy_report.R
\name{accuracy_report}
\alias{accuracy_report}
\title{Produce table of accuracy and area measures}
\usage{
accuracy_report(err_mat, acc_mat, classes, digits = 1, z = 1.96)
}
\arguments{
\item{err_mat}{The error matrix, with reference sample and map sample in rows}

\item{acc_mat}{The output list produced by `accuracy`}

\item{classes}{Vector of names for the map/reference classes}

\item{digits}{Number of decimal values in outputs}

\item{z}{Width of confidence interval for reported standard errors}
}
\value{
A character matrix.
}
\description{
Produce table of accuracy and area measures
}
\details{
Produces a matrix for reporting that contains the
configuration for reporting map accuracy recommended by Stehman and Foody
(2019), with adjusted areas of each class provided.
}
\examples{
# Example 1: Table 4 in Olofsson et al 2014a
err_mat <- cbind(r1 = c(97, 3, 2), r2 = c(0, 279, 1), r3 = c(3, 18, 97))
areas <- c(22353, 1122543, 610228)
acc_mat <- accuracy(err_mat, areas)
accuracy_report(err_mat, acc_mat, classes = paste0("Class", 1:3))
}
\references{
Stehman, S.V. & Foody, G.M. (2019) Key issues in rigorous accuracy assessment
of land cover products. Remote Sensing of Environment, 231, 111199.
}
