% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accuracy_reports}
\alias{accuracy_reports}
\title{Model performance metrics per iteration}
\description{
Results of accuracy assessment performed with map reference
sample, including margins of error for overall, producer's, and user's
accuracy, as well as class area estimates, with confidence intervals.
}
\examples{
data(accuracy_reports)

}
\keyword{data}
