% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_map_functions.R
\name{assignment_history}
\alias{assignment_history}
\title{Collect assignment history for a given worker}
\usage{
assignment_history(assignment_data, accuracy_data, workerid)
}
\arguments{
\item{assignment_data}{Assignment data table}

\item{accuracy_data}{Accuracy data table}

\item{workerid}{Integer ID of worker}
}
\value{
Tibble of measurements
}
\description{
Collect assignment history for a given worker
}
\keyword{internal}
