% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{consensus_high_low}
\alias{consensus_high_low}
\title{Model performance metrics per iteration}
\description{
Performance scores from final iteration for 4 AOIs (1, 2, 8, 15)
where label strategy experiments were run: models trained with consensus
labels versus those trained with most and least accurate individual labels
}
\examples{
data(consensus_high_low)
}
\keyword{data}
