% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_map_functions.R
\name{consensus_map_creation2}
\alias{consensus_map_creation2}
\title{Create consensus labels and calculated Bayesian measures of uncertainty}
\usage{
consensus_map_creation2(
  kml_data,
  hit_data,
  assignment_data,
  accuracy_data,
  qual_assignment_data,
  qual_accuracy_data,
  user_maps,
  ref_grid,
  kmlid,
  mode,
  diam,
  risk_threshold,
  categories,
  category = "field",
  unsure_category = "unsure1",
  qsite = FALSE,
  crs = "4326"
)
}
\arguments{
\item{kml_data}{kml_data (see notes for type)}

\item{hit_data}{hit_data (see notes for type)}

\item{assignment_data}{assignment_data (see notes for type)}

\item{accuracy_data}{accuracy_data(see notes for type)}

\item{qual_assignment_data}{qual_assignment_data (see notes for type)}

\item{user_maps}{Either an sf object holding maps or postgresql connection}

\item{ref_grid}{Primary reference grid}

\item{kmlid}{Name of label site (must match name in ref_grid)}

\item{mode}{Label-making method: 'consensus', 'high' or 'low'. The first
creates a merged label, 'high' or 'low' uses the highest or the lowest
scoreing workers maps.}

\item{diam}{Half-width of cell in reference grid}

\item{risk_threshold}{Value from 0-1 indicating threhold for flagging pixels
with high label uncertainty}

\item{categories}{Table holding class categories}

\item{category}{Single class category to process. Default if "field"}

\item{unsure_category}{Name of class category indicating uncertainty}

\item{qsite}{Q ir F site?  Default is FALSE}

\item{crs}{Coordinate reference system, current 4326, otherwise proj4string}
}
\value{
List containing rasters (label, heat map, risk map) and Bayesian
measures of uncertainty.
}
\description{
Create consensus labels and calculated Bayesian measures of uncertainty
}
\details{
This function is a modification of the function in rmapaccuracy,
which is designed to allowed consensus labels to be made as a batch.
}
