% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{error_matrices}
\alias{error_matrices}
\title{Error matrices using map reference sample}
\description{
Error matrices are calculated for the per pixel and segmented
field boundary map, for particular zones consisted of grouped AOIs and AEZ,
and for all of Ghana.
}
\examples{
data(error_matrices)
}
\keyword{data}
