% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_map_functions.R
\name{get_db_tables}
\alias{get_db_tables}
\title{Get pointers to key database tables from labeller}
\usage{
get_db_tables(connection, dbtables = "default")
}
\arguments{
\item{connection}{postgresql connection to instance database}

\item{dbtables}{Vector of table names to get. Currently "default" brings up
names of tables needed for consensus labelling}
}
\value{
Named list of lazy connections to labeller database tables
}
\description{
Get pointers to key database tables from labeller
}
