% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_map_functions.R
\name{get_user_maps}
\alias{get_user_maps}
\title{Get user_maps for generating consensus labels}
\usage{
get_user_maps(assignmentid, categories, category, grid_poly, user_maps, qsite)
}
\arguments{
\item{assignmentid}{ID of assignment for which to get maps}

\item{categories}{Table holding class categories}

\item{category}{Single class category of interest}

\item{grid_poly}{Polygon of the reference grid corresponding to assignment}

\item{user_maps}{Either an sf object with fields or a postresql connection}

\item{qsite}{TRUE or FALSE, passed through from consensus_map_creation}
}
\value{
An sf geometry object, empty or containing buffered/unioned
boundaries clipped to the grid
}
\description{
Get user_maps for generating consensus labels
}
\keyword{internal}
