% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{image_quality}
\alias{image_quality}
\title{Model performance metrics per iteration}
\description{
Quality scores for the seasonal image composites
}
\examples{
data(image_quality)
}
\keyword{data}
