% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iteration_metrics}
\alias{iteration_metrics}
\title{Model performance metrics per iteration}
\description{
Performance scores per iteration for each of the AOIs, including
score changes measures between iterations
}
\examples{
data(iteration_metrics)
}
\keyword{data}
