% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{label_risk}
\alias{label_risk}
\title{Bayesian Risk metrics per site with summary statistics}
\description{
Provides the mean Bayesian Risk and proportion of risky pixels
for each site per AOI, and summarized per AOI.
}
\examples{
data(label_risk)
}
\keyword{data}
