% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{label_risk_metrics}
\alias{label_risk_metrics}
\title{Summary Bayesian Risk metrics along with per AOI model performance metrics}
\description{
Provides the mean Bayesian Risk per AOI and scores for various
Random Forests performance metrics assessed against validation sites.
}
\examples{
data(label_risk_metrics)
}
\keyword{data}
