% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeller_data.R
\name{labeller_connection}
\alias{labeller_connection}
\title{Connection to database}
\usage{
labeller_connection(host, dbname, user, password)
}
\arguments{
\item{host}{URL of host}

\item{dbname}{Database name}

\item{user}{Database user name}

\item{password}{Database password}
}
\description{
Connection to database
}
