% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeller_data.R
\name{labeller_data}
\alias{labeller_data}
\title{Read in data from multiple labeller databases}
\usage{
labeller_data(connection, instances, table, variables, as_tibble = TRUE)
}
\arguments{
\item{connection}{Database connection provided by labeller_connection}

\item{instances}{A character vector specifying the complete URLs of servers}

\item{table}{Table names to get from each instance}

\item{variables}{Vector of variable names to get from the table}

\item{as_tibble}{Returns output as single tibble (TRUE) or list (FALSE)}
}
\description{
Read in data from multiple labeller databases
}
