% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapped_areas}
\alias{mapped_areas}
\title{Croplands areas}
\description{
Cropland areas calculated from the classified cropland
probabilities and from the segmented field boundaries per AOI. Areas
per AEZ are also calculated for the classified cropland
probabilities
}
\examples{
data(mapped_areas)
}
\keyword{data}
