% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_map_functions.R
\name{qual_assignment_history}
\alias{qual_assignment_history}
\title{Collect qualification assignment history for a given worker}
\usage{
qual_assignment_history(qual_assignment_data, qual_accuracy_data, workerid)
}
\arguments{
\item{qual_assignment_data}{Assignment data table}

\item{qual_accuracy_data}{Qualification accuracy table}

\item{workerid}{Integer ID of worker}
}
\value{
Tibble of measurements
}
\description{
Collect qualification assignment history for a given worker
}
\keyword{internal}
