% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tile_key}
\alias{tile_key}
\title{Training and validation sites for AOIs 1-16}
\description{
Tibble cross referencing AOIs, tile grid cell IDs, and col, row
numbers used for prediction tiles
}
\examples{
data(tile_key)
}
\keyword{data}
