% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{train_val_sites}
\alias{train_val_sites}
\title{Training and validation sites for AOIs 1-16}
\description{
Centroids of all points used for training and validation,
excluding run 0 for AOI 3.
}
\examples{
data(train_val_sites)
}
\keyword{data}
