% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_map_functions.R
\name{write_consensus_results}
\alias{write_consensus_results}
\title{Write consensus results to instance database and s3}
\usage{
write_consensus_results(
  results,
  params,
  output_riskmap,
  s3bucket = NULL,
  dbupdate = TRUE,
  coninfo = NULL
)
}
\arguments{
\item{results}{Output list from consensus_map_creation}

\item{params}{Parameter file read in from common/}

\item{output_riskmap}{Save output risk map?}

\item{s3bucket}{Name of s3bucket, read from db}

\item{dbupdate}{Update the database with risk results or not. Default = TRUE}

\item{coninfo}{Database connection information from mapper_connect()}
}
\description{
Write consensus results to instance database and s3
}
\details{
This function will not save a consensus map to s3 if s3bucket is
kept as null, useful for testing. The same applies for the risk map
}
