% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_utils.R
\name{xys_from_rowcol}
\alias{xys_from_rowcol}
\title{Get center coordinates from row column for learner's grid reference system}
\usage{
xys_from_rowcol(r, c, e = c(-17.541, 51.459, -35.46, 37.54), res = 0.05)
}
\arguments{
\item{r}{Vector of column indices to refer to main grid}

\item{c}{Vector of row indices to refer to main grid}

\item{e}{Extent of main grid, defaults to current one used for Africa}

\item{res}{Resolution of main grid, defaults to 0.5}
}
\value{
A matrix of
}
\description{
Get center coordinates from row column for learner's grid reference system
}
